Make RetriableRequest work with multiple versions of Go. (#154)
Moved version-specific code to source files with appropriate build tags
and added earlier versions of Go to travis CI.
Exclude the unixtime tests from 1.6 as they rely on functionality that
isn't available in 1.6 (t.Run).
