+++ .travis.yml
@@ -4,6 +4,8 @@ language: go
 
 go: 
   - 1.8
+  - 1.7
+  - 1.6
 
 install:
   - go get -u github.com/golang/lint/golint
