+++ autorest/date/unixtime_test.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 package date
 
 import (
